Moved get_and_cleanup_actions to the resource object.
